feat(gateway): cors config 작성
