+++ gateway-service/src/main/java/com/example/gatewayservice/common/security/CorsConfig.java
@@ -1,0 +1,43 @@
+package com.example.gatewayservice.common.security;
+
+
+import jakarta.annotation.PostConstruct;
+import java.util.Arrays;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.reactive.CorsWebFilter;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
+
+import java.util.List;
+
+@Configuration
+public class CorsConfig {
+    @Value("${cors.allowed.origin}")
+    private String allowedOrigin;
+
+    private List<String> allowedOriginList;
+
+
+
+    @PostConstruct
+    public void init() {
+
+        allowedOriginList = Arrays.asList(allowedOrigin.split(","));
+
+    }
+    @Bean
+    public CorsWebFilter corsWebFilter() {
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(allowedOriginList); // 허용할 프론트 URL
+        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        config.setAllowedHeaders(List.of("*"));
+        config.setAllowCredentials(true);
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", config);
+
+        return new CorsWebFilter(source);
+    }
+}
